+++ src/Http/AdminApi/Authentication/Contract/AuthenticatedHttpClientInterface.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\Authentication\Contract;
+
+use Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\Authentication\Exception\AuthenticationFailed;
+use Psr\Http\Client\ClientExceptionInterface;
+use Psr\Http\Client\ClientInterface;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
+
+interface AuthenticatedHttpClientInterface extends ClientInterface
+{
+    /**
+     * @throws AuthenticationFailed     if an error happens while authenticating the request
+     * @throws ClientExceptionInterface if an error happens while processing the request
+     */
+    public function sendRequest(RequestInterface $request): ResponseInterface;
+}